src/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
openpyxl
python-dotenv
PyYAML
plotly
plotly-express
openai
trainsformers
sentence-transformers
huggingface_hub
langchain-text-splitters
accelerate
bitsandbytes 
datasets
pytorch
umap-learn
wandb
evaluate
deepeval
weave 
trulens 
trulens-eval
crewai

# for loRA fine tuning 
peft
trl

# for audio 
librosa